requirements.txt
absl-py==1.4.0
accelerate==0.33.0
altair==4.2.2
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_record==0.5.1
astunparse==1.6.3
attrs==24.2.0
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
blis==0.7.11
build==1.2.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
cloud-tpu-client==0.10
cloudpathlib==0.19.0
cloudpickle==3.0.0
confection==0.1.5
contourpy==1.3.0
cryptography==3.4.8
cycler==0.12.1
cymem==2.0.8
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distrax==0.1.5
distro==1.7.0
dm-tree==0.1.8
docstring_parser==0.16
dopamine_rl==4.0.9
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
entrypoints==0.4
etils==1.7.0
exceptiongroup==1.2.2
fastai==2.7.17
fastcore==1.7.1
fastdownload==0.0.7
fastjsonschema==2.20.0
fastprogress==1.0.3
filelock==3.15.4
flatbuffers==24.3.25
flax==0.8.4
fonttools==4.53.1
fsspec==2024.6.1
funcsigs==1.0.2
gast==0.6.0
GDAL==3.6.4
gdown==5.1.0
gensim==4.3.3
gin-config==0.5.0
google==2.0.3
google-api-core==1.34.1
google-api-python-client==1.8.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.8.0
google-colab @ file:///colabtools/dist/google_colab-1.0.0.tar.gz#sha256=dd52203d74730c391700de4cebd4c46a36d1d67482bd0f068d6a925885510841
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
gspread==6.0.2
gspread-dataframe==3.3.1
gym==0.25.2
gym-notices==0.0.8
h5py==3.11.0
httplib2==0.22.0
huggingface-hub==0.24.6
idna==3.8
imageio==2.35.1
immutabledict==4.2.0
importlib-metadata==4.6.4
importlib_resources==6.4.4
iniconfig==2.0.0
ipykernel==5.5.6
ipyparallel==8.8.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jax==0.4.26
jaxlib==0.4.26
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==6.1.12
jupyter-console==6.1.0
jupyter-server==1.24.0
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kaggle==1.6.17
kagglehub==0.2.9
keras==2.15.0
keyring==23.5.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
libtpu-nightly==0.1.dev20240403+default
llvmlite==0.43.0
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.10
nbclassic==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==6.5.5
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
optax==0.1.9
orbax-checkpoint==0.4.4
packaging==24.1
pandas==2.1.4
pandas-stubs==2.1.4.231227
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip-tools==7.4.1
platformdirs==4.2.2
plotly==5.15.0
pluggy==1.5.0
pooch==1.8.2
portpicker==1.5.2
preshed==3.0.9
prometheus_client==0.20.0
promise==2.3
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.4
pyproject_hooks==1.1.0
PySocks==1.7.1
pytest==7.4.4
python-apt==2.4.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
pyzmq==24.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requirements-parser==0.9.0
rich==13.8.0
rpds-py==0.20.0
rpy2==3.4.2
rsa==4.9
safetensors==0.4.4
scikit-image==0.23.2
scikit-learn==1.3.2
scipy==1.13.1
seaborn==0.13.1
SecretStorage==3.3.1
Send2Trash==1.8.3
sentencepiece==0.1.99
shellingham==1.5.4
simple_parsing==0.1.5
six==1.16.0
sklearn-pandas==2.2.0
smart-open==7.0.4
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
StrEnum==0.4.15
sympy==1.13.2
tenacity==9.0.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow @ https://storage.googleapis.com/cloud-tpu-tpuvm-artifacts/tensorflow/tf-2.15.0/tensorflow-2.15.0-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl#sha256=23069418d5bea74d96f14f0ee47e8f429f0da4267df91769040a80f477e6ee05
tensorflow-datasets==4.9.6
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.15.0
tensorflow-probability==0.24.0
tensorflow-text==2.15.0
tensorstore==0.1.45
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
tf-slim==1.1.0
tf_keras==2.15.1
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.8.28
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch @ https://download.pytorch.org/whl/cpu/torch-2.4.0%2Bcpu-cp310-cp310-linux_x86_64.whl#sha256=0e59377b27823dda6d26528febb7ca06fc5b77816eaa58b4420cc8785e33d4ce
torch-xla @ https://storage.googleapis.com/pytorch-xla-releases/wheels/tpuvm/torch_xla-2.4.0+libtpu-cp310-cp310-manylinux_2_28_x86_64.whl#sha256=d973d6c0bd0cda884a9a68778b2d3a2d3cc4f2876dd21e5c95f93091904aab19
torchaudio @ https://download.pytorch.org/whl/cpu/torchaudio-2.4.0%2Bcpu-cp310-cp310-linux_x86_64.whl#sha256=b72ce53071b0bcd4c52b81c8ef4fbc01649cc8dacc646982cf9ef59ac298ff91
torchvision @ https://download.pytorch.org/whl/cpu/torchvision-0.19.0%2Bcpu-cp310-cp310-linux_x86_64.whl#sha256=7b2b46d3c757fa1fe7d08d48bdae6c9d97d82ace707474a79219c10991fee6ff
tornado==6.3.3
tpu-info @ git+https://github.com/google/cloud-accelerator-diagnostics/@e898c785f1b1aa7daab5484aefa05a7fad9d32e9#subdirectory=tpu_info
tqdm==4.66.5
traitlets==5.7.1
transformers==4.44.2
trax==1.4.1
typer==0.12.5
types-pytz==2024.1.0.20240417
types-setuptools==74.0.0.20240831
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==3.0.1
urllib3==2.0.7
vega-datasets==0.9.0
wadllib==1.3.6
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
widgetsnbextension==3.6.8
wrapt==1.14.1
zipp==3.20.1
